requirements.txt
flask
flask-login
google-cloud-firestore
google-cloud-storage
joblib
scikit-learn
pandas
prophet